# docs/swagger.yaml
basePath: /
definitions:
  action.CopyMoveFileRequest:
    properties:
      onConflict:
        allOf:
        - $ref: '#/definitions/action.OnConflictAction'
        description: OnConflict is the action to take when the file already exists
      sourcePath:
        description: SourcePath is the path of the file to copy
        type: string
      targetPath:
        description: TargetPath is the path where the file will be copied
        type: string
    required:
    - onConflict
    - sourcePath
    - targetPath
    type: object
...
  action.OnConflictAction:
    enum:
    - overwrite
    - abort
    type: string
    x-enum-varnames:
    - OnConflictActionOverwrite
    - OnConflictActionAbort
...
  onlyoffice.onlyOfficeUpdateRequest:
    properties:
      status:
        description: 2 for update
        type: integer
      url:
        description: The URL of the file
        type: string
    type: object
...
info:
  contact:
    email: vvbbnn00@gmail.com
    name: vvbbnn00
    url: https://github.com/vvbbnn00/goflet
  description: Goflet is a lightweight file upload and download service written in
    Go.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Goflet API
  version: unknown
paths:
  /api/action/copy:
    post:
      consumes:
      - application/json
      description: Copy a file from one location to another, if you want to move a
        file, use the move action instead.
      parameters:
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/action.CopyMoveFileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "404":
          description: File not found
          schema:
            type: string
        "409":
          description: File exists
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - Authorization: []
      summary: Copy File
      tags:
      - Action
...
  /api/action/move:
    post:
      consumes:
      - application/json
      description: Move a file from one location to another, the performance of moving
        is better than copying.
      parameters:
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/action.CopyMoveFileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "404":
          description: File not found
          schema:
            type: string
        "409":
          description: File exists
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - Authorization: []
      summary: Move File
      tags:
      - Action
  /api/image/{path}:
    get:
      description: Get processed image, {path} should be the relative path of the
        file, starting from the root directory, e.g. /image/path/to/image.jpg
      parameters:
      - description: File path
        in: path
        name: path
        required: true
        type: string
      - description: Width
        in: query
        name: w
        type: integer
      - description: Height
        in: query
        name: h
        type: integer
      - description: Quality, 0-100
        in: query
        name: q
        type: integer
      - description: Format
        enum:
        - jpg
        - png
        - gif
        in: query
        name: f
        type: string
      - description: Angle, 0-360
        in: query
        name: a
        type: integer
      - description: Scale type
        enum:
        - fit
        - fill
        - resize
        - fit_width
        - fit_height
        in: query
        name: s
        type: string
      produces:
      - image/jpeg
      - ' image/png'
      - ' image/gif'
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "404":
          description: File not found
          schema:
            type: string
        "413":
          description: File too large
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - Authorization: []
      summary: Get Image
      tags:
      - Image
...
  /api/onlyoffice/{path}:
    post:
      consumes:
      - application/json
      description: OnlyOffice callback for file updates, when the status is 2, the
        file has been updated, {path} should be the relative path of the file, starting
        from the root directory, e.g. /onlyoffice/path/to/file.txt
      parameters:
      - description: File path
        in: path
        name: path
        required: true
        type: string
      - description: OnlyOffice update request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/onlyoffice.onlyOfficeUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "404":
          description: File not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - Authorization: []
      summary: OnlyOffice Callback
      tags:
      - OnlyOffice
...
securityDefinitions:
  Authorization:
    description: 'You need to provide a valid jwt token in the header, in headers,
      you should provide a key-value pair like this: Authorization: Bearer xxxxxx;
      The token has the same effect as the token in the query string, but it is more
      secure than the token in the query string. Or you can just provide the token
      in the query string, like this: ?token=xxxxxx. More info about jwt: https://github.com/vvbbnn00/goflet?tab=readme-ov-file#authentication-method'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"